refactor: move PersonalDetailsPreview component to its own file
